Move DTS management to separate client type.
